JWT refresh add

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 class Item(Resource):
@@ -18,6 +18,7 @@ class Item(Resource):
         return {"message": "Item not found"},404
 
     #POST
+    @jwt_required(fresh=True)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {"message": "An item with name '{}' already exist".format(name)},400
@@ -34,7 +35,11 @@ class Item(Resource):
         return item.json(), 201
     
     #DELETE
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required"},401
         item = ItemModel.find_by_name(name)
         if item :
             ItemModel.delete_from_db()
@@ -57,6 +62,15 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True) #different things if is loged or not
     def get(self):
-        return {"items": [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items =  [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items" : items},200
+
+        return {
+            "items" : [item["name"] for item in items],
+            "message" : "More data available if you log in"
+        },200
         #return { "items": list(map(lambda x: x.json(), ItemModel.query.all()))}
